Handle status code 204, ignore smart playlists.
Newer versions of the Plex Media Server return HTTP status code 204 when deleting a playlist. This causes PlexAPI versions 4.1.2 and lower to throw a BadRequest exception. We now check the PlexAPI version to see if we need to work around this issue or not. This was fixed in the PlexAPI with https://github.com/pkkid/python-plexapi/pull/580, so we will assume that a version higher than 4.1.2 will contain this fix.